pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gpt-api-wrapper"
version = "0.2.0"
description = "Simple wrapper around multiple LLM providers with single and batch operations plus conversation history support"
readme = "README.md"
authors = [{name = "Sanps"}]
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.34.0",
    "xai-sdk>=0.2.0",
    "google-genai>=0.3.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools]
packages = ["gpt_api_wrapper", "gpt_api_wrapper.providers"]
